refactor(cli): reorganize modules into api/ and lib/ directories
