+++ apps/cli/src/api/agent-registry.ts
