+++ apps/cli/src/lib/a2a.ts
